// src/src/features/Home/types.ts
export interface ICurrencyData {
    name: string,
    title: string,
    price: string,
    time: string,
    rateOfChange: number
}